Moved common_commands.py to __init__.py
